Se realizo plantilla para la interfaz IprogramaService

+++ src/main/java/com/example/programa_team1/model/entities/Programa.java
@@ -13,7 +13,7 @@ public class Programa {
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
 
-    private int codPrograma, codFacultad;
+    private long codPrograma, codFacultad;
     private String nomPrograma, nomFacultad;
